refactor: separate handler and validator tests add extension methods
