+++ src/Content/Backend/Solution/Monaco.Template.Backend.Application.Tests/Features/Country/GetCountryListTests.cs
@@ -5,10 +5,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.Extensions.Primitives;
-using MockQueryable.Moq;
 using Monaco.Template.Backend.Application.DTOs;
 using Monaco.Template.Backend.Application.Features.Country;
 using Monaco.Template.Backend.Application.Infrastructure.Context;
+using Monaco.Template.Backend.Common.Tests;
 using Monaco.Template.Backend.Common.Tests.Factories;
 using Moq;
 using Xunit;
@@ -16,17 +16,20 @@ using Xunit;
 namespace Monaco.Template.Backend.Application.Tests.Features.Country;
 
 [ExcludeFromCodeCoverage]
-[Trait("Application Queries", "Country Queries")]
+[Trait("Application Queries", "Get Country List")]
 public class GetCountryListTests
 {
+	private readonly Mock<AppDbContext> _dbContextMock = new();
+
 	[Theory(DisplayName = "Get country list without params succeeds")]
 	[AnonymousData]
 	public async Task GetCountryListWithoutParamsSucceeds(List<Domain.Model.Country> countries)
 	{
-		var dbContextMock = SetupMock(countries);
+		_dbContextMock.CreateAndSetupDbSetMock(countries);
+
 		var query = new GetCountryList.Query(new List<KeyValuePair<string, StringValues>>());
 
-		var sut = new GetCountryList.Handler(dbContextMock.Object);
+		var sut = new GetCountryList.Handler(_dbContextMock.Object);
 		var result = await sut.Handle(query, new CancellationToken());
 
 		result.Should()
@@ -39,7 +42,7 @@ public class GetCountryListTests
 	[AnonymousData]
 	public async Task GetCountryListWithParamsSucceeds(List<Domain.Model.Country> countries)
 	{
-		var dbContextMock = SetupMock(countries);
+		_dbContextMock.CreateAndSetupDbSetMock(countries);
 		var countriesSet = countries.GetRange(0, 2);
 
 		var queryString = new List<KeyValuePair<string, StringValues>>
@@ -51,7 +54,7 @@ public class GetCountryListTests
 
 		var query = new GetCountryList.Query(queryString);
 
-		var sut = new GetCountryList.Handler(dbContextMock.Object);
+		var sut = new GetCountryList.Handler(_dbContextMock.Object);
 
 		var result = await sut.Handle(query, new CancellationToken());
 
@@ -59,15 +62,5 @@ public class GetCountryListTests
 			  .HaveCount(countriesSet.Count).And
 			  .Contain(x => countriesSet.Any(c => c.Name == x.Name)).And
 			  .BeInDescendingOrder(x => x.Name);
-	}
-
-	private static Mock<AppDbContext> SetupMock(IEnumerable<Domain.Model.Country> countries)
-	{
-		var dbSetMock = countries.AsQueryable().BuildMockDbSet();
-		var dbContextMock = new Mock<AppDbContext>();
-		dbContextMock.Setup(x => x.Set<Domain.Model.Country>())
-					 .Returns(dbSetMock.Object);
-
-		return dbContextMock;
 	}
 }
